Rename cl text input file
